requirements.txt
notebook
ipywidgets
tqdm
python-dotenv
requests
openai
langchain
Flask
transformers
torch